# 8332602: [s390x] Improve itable_stub