--- a/src/hotspot/cpu/s390/macroAssembler_s390.hpp
+++ b/src/hotspot/cpu/s390/macroAssembler_s390.hpp
@@ -670,12 +670,12 @@ class MacroAssembler: public Assembler {
                                Label&             no_such_interface,
                                bool               return_method = true);
 
-  void lookup_interface_method_stub(Register r_recv_klass,
-                                    Register r_holder_klass,
-                                    Register r_resolved_klass,
-                                    Register r_method_result,
-                                    Register r_temp,
-                                    Register r_temp2,
+  void lookup_interface_method_stub(Register recv_klass,
+                                    Register holder_klass,
+                                    Register resolved_klass,
+                                    Register method_result,
+                                    Register temp,
+                                    Register temp2,
                                     int      itable_index,
                                     Label&   nl_no_such_interface);