s390x Port similar to [JDK-8305959 (x86)](https://bugs.openjdk.org/browse/JDK-8305959) and [JDK-8307352(aarch64)](https://bugs.openjdk.org/browse/JDK-8307352)

Testing: I ran `tier1` test on fastdebug & release VM; I didn't see any regression there;

Benchmarking: 
```
Without Patch: 
Benchmark                        Mode  Cnt   Score    Error  Units
InterfaceCalls.test1stInt2Types  avgt   12   1.924 ±  0.001  ns/op
InterfaceCalls.test1stInt3Types  avgt   12  13.925 ±  0.014  ns/op
InterfaceCalls.test1stInt5Types  avgt   12  16.591 ±  0.045  ns/op
InterfaceCalls.test2ndInt2Types  avgt   12   2.028 ±  0.013  ns/op
InterfaceCalls.test2ndInt3Types  avgt   12   7.634 ±  0.049  ns/op
InterfaceCalls.test2ndInt5Types  avgt   12  16.231 ±  1.222  ns/op
InterfaceCalls.testIfaceCall     avgt   12  16.587 ±  0.058  ns/op
InterfaceCalls.testIfaceExtCall  avgt   12  17.532 ±  0.024  ns/op
InterfaceCalls.testMonomorphic   avgt   12   0.746 ±  0.001  ns/op
Finished running test 'micro:vm.compiler.InterfaceCalls'


With Patch: 

Benchmark                        Mode  Cnt   Score    Error  Units
InterfaceCalls.test1stInt2Types  avgt   12   1.929 ±  0.012  ns/op
InterfaceCalls.test1stInt3Types  avgt   12  13.280 ±  0.093  ns/op
InterfaceCalls.test1stInt5Types  avgt   12  16.169 ±  0.364  ns/op
InterfaceCalls.test2ndInt2Types  avgt   12   6.758 ±  4.473  ns/op
InterfaceCalls.test2ndInt3Types  avgt   12  11.772 ±  2.411  ns/op
InterfaceCalls.test2ndInt5Types  avgt   12  15.099 ±  0.081  ns/op
InterfaceCalls.testIfaceCall     avgt   12  15.972 ±  0.021  ns/op
InterfaceCalls.testIfaceExtCall  avgt   12  16.600 ±  0.322  ns/op
InterfaceCalls.testMonomorphic   avgt   12   0.746 ±  0.001  ns/op
Finished running test 'micro:vm.compiler.InterfaceCalls'
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [ ] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8332602](https://bugs.openjdk.org/browse/JDK-8332602): [s390x] Improve itable_stub (**Enhancement** - P4)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/19698/head:pull/19698` \
`$ git checkout pull/19698`

Update a local copy of the PR: \
`$ git checkout pull/19698` \
`$ git pull https://git.openjdk.org/jdk.git pull/19698/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 19698`

View PR using the GUI difftool: \
`$ git pr show -t 19698`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/19698.diff">https://git.openjdk.org/jdk/pull/19698.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/19698#issuecomment-2165648765)